# openapi.yaml
openapi: 3.1.0
info:
  title: SupKube Backend API
  version: "0.9.1.10-alpha"
  description: |
    SupKube 后端 REST API 的机器可读契约。

    **本文件状态**：cross-cutting 契约（认证 / 错误信封 / `/status` / 列表信封 / RBAC）
    已完整定义；resource paths 以**已核对的核心资源**为种子，按模块增量补全。
    **完整端点目录（13 组 ~90 端点）以 [API-REFERENCE.md](API-REFERENCE.md) §7 为准**，
    其权威源是 `supkube-backend/internal/auth/rbac.go` 的 `permissionTable`。

    扩展字段 `x-min-role`：调用该操作所需最低角色（viewer < editor < admin）。
    带 `x-namespace-scoped: true` 的 editor 操作还要过 namespace 作用域检查。
  contact:
    name: SupKube
    url: https://charts.supkube.com
  license:
    name: Proprietary
servers:
  - url: https://{host}/api/v1
    description: 外部（ingress）
    variables:
      host:
        default: supkube.example.com
  - url: http://supkube-backend.{namespace}.svc:8080/api/v1
    description: 集群内
    variables:
      namespace:
        default: supkube
...
security:
  - bearerAuth: []
  - basicAuth: []
paths:
...
  /import-policies:
    get:
      tags: [import-policies]
      operationId: listImportPolicies
      summary: 导入策略列表
      x-min-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ImportPolicy" }
    post:
      tags: [import-policies]
      operationId: createImportPolicy
      summary: 创建导入策略（替代 Velero backupSyncPeriod 60s 兜底）
      x-min-role: editor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ImportPolicy" }
      responses:
        "201": { description: Created }
        "400":
          description: 校验失败（ERR_IMPORTPOLICY_*）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StructuredError" }
        "409":
          description: 同名已存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StructuredError" }
              example:
                error: "import policy 'foo' already exists"
                code: ERR_IMPORTPOLICY_ALREADY_EXISTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`。token 可为静态 API Token（脚本/MCP/CI）
        或 OIDC JWT（人类经 Dex 登录）。后端按内容自动分派。
    basicAuth:
      type: http
      scheme: basic
      description: htpasswd 兜底（无 IdP 的小集群 / air-gap）。
...
  schemas:
...
    StructuredError:
      type: object
      description: 结构化错误信封（ADR-035，新模块）
      required: [error, code]
      properties:
        error: { type: string, description: 给人看的话 }
        code:
          type: string
          description: ERR_<DOMAIN>_<REASON>，稳定不随文案变
          example: ERR_IMPORTPOLICY_ALREADY_EXISTS
...
    ImportPolicy:
      type: object
      description: |
        导入策略（PRD-009 / ADR-038）。字段为当前已核对子集，完整以
        internal/importpolicy/types.go 为准。
      properties:
        name:
          type: string
          description: DNS-1123 label
        mode:
          type: string
          enum: [Continuous, Scheduled]
          description: Continuous=time.Ticker；Scheduled=robfig/cron
        bsl:
          type: string
          description: 源 BackupStorageLocation 名
        schedule:
          type: string
          description: cron 表达式（Scheduled 模式）
        intervalSeconds:
          type: integer
          description: Continuous 模式间隔（有下限，过短报 ERR_IMPORTPOLICY_INTERVAL_TOO_SHORT）
        fingerprintMode:
          type: string
          enum: [enforce, warn, disabled]
        lastError:
          type: string
          readOnly: true